Time: 0 ms (100.00%) | Memory: 8 MB (42.11%) - LeetSync

+++ 278-first-bad-version/first-bad-version.cpp
@@ -4,10 +4,10 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        int start =1, end= n;
+        long long start =1, end= n;
 
         while(start<end){
-            int mid =  start+(end-start)/2;
+            long long mid =  start+(end-start)/2;
             if(isBadVersion(mid))
             end=mid;
             else
